requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
supabase>=1.0.0
python-dotenv>=0.19.0
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.0.0 